+++ sprint8/asyncHandlerFunction.js
@@ -1,0 +1,26 @@
+import { Prisma } from "@prisma/client";
+
+// 에러 처리 함수
+function asyncHandler(handler) {
+  return async function (req, res) {
+    try {
+      await handler(req, res);
+    } catch (e) {
+      if (
+        e.name === 'StructureError' ||
+        e instanceof Prisma.PrismaClientValidationError
+      ) {
+        res.status(400).send({ message: e.message });
+      } else if (
+        e instanceof Prisma.PrismaClientKnownRequestError &&
+        e.code === 'P2025'
+      ) {
+        res.status(404).send({ message: 'Cannot find given id.' });
+      } else {
+        res.status(500).send({ message: e.message });
+      }
+    }
+  }
+}
+
+export default asyncHandler;
